+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import Home from "./Pages/Home";
-import Login from "./Pages/Login";
+import Home from "./Pages/Home/Home";
+import Login from "./Pages/Login/Login";
