+++ src/Store/store.js
@@ -1,9 +1,13 @@
 import { applyMiddleware, compose, createStore,combineReducers } from "redux";
+import { takeEvery} from "@redux-saga/core/effects";
 import createSagaMiddleware from "@redux-saga/core";
-import rootSaga from "../Saga/Saga";
-//import nameReducer from "./Reducers/UserReduceer";
-import reducerFunction from "./Reducers/CardReducer";
+import reducerFunction from "../Pages/Home/cardReducer";
+import { getCards, saveCard } from "../Pages/Home/homeSaga";
 
+function* rootSaga() {
+  yield takeEvery('addNew', saveCard);
+  yield takeEvery('getCards', getCards)
+}
 
 const sagaMiddleware = createSagaMiddleware();
