+++ src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { Provider } from 'react-redux';
-import store from '../src/Store/index'
+import store from '../src/Store/store'
 
 ReactDOM.render(
   <React.StrictMode>
